requirements.txt
asgiref==3.2.3
attrs==19.3.0
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
coveralls==1.10.0
Django==3.0.3
djangorestframework==3.11.0
docopt==0.6.2
idna==2.8
importlib-metadata==1.5.0
more-itertools==8.2.0
packaging==20.1
phonenumbers==8.11.3
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
py3dns==3.2.1
pycodestyle==2.5.0
pyparsing==2.4.6
pytest==5.3.5
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
validate-email==1.3
wcwidth==0.1.8
zipp==2.1.0
